+++ microsphere-i18n-spring/src/main/java/io/microsphere/i18n/spring/annotation/I18nImportBeanDefinitionRegistrar.java
@@ -18,11 +18,11 @@ package io.microsphere.i18n.spring.annotation;
 
 import io.microsphere.i18n.spring.DelegatingServiceMessageSource;
 import io.microsphere.i18n.spring.beans.factory.ServiceMessageSourceFactoryBean;
-import io.microsphere.i18n.spring.beans.factory.config.AcceptHeaderLocaleResolverBeanPostProcessor;
-import io.microsphere.i18n.spring.beans.factory.config.I18nLocalValidatorFactoryBeanPostProcessor;
 import io.microsphere.i18n.spring.beans.factory.support.ServiceMessageSourceBeanLifecyclePostProcessor;
 import io.microsphere.i18n.spring.context.I18nApplicationListener;
 import io.microsphere.i18n.spring.context.MessageSourceAdapter;
+import io.microsphere.i18n.spring.validation.beanvalidation.I18nLocalValidatorFactoryBeanPostProcessor;
+import io.microsphere.i18n.spring.web.servlet.AcceptHeaderLocaleResolverBeanPostProcessor;
 import io.microsphere.logging.Logger;
 import org.springframework.beans.factory.config.BeanDefinition;
 import org.springframework.beans.factory.support.BeanDefinitionRegistry;
@@ -41,9 +41,12 @@ import static io.microsphere.i18n.spring.constants.I18nConstants.DEFAULT_ENABLED
 import static io.microsphere.i18n.spring.constants.I18nConstants.ENABLED_PROPERTY_NAME;
 import static io.microsphere.i18n.spring.constants.I18nConstants.SERVICE_MESSAGE_SOURCE_BEAN_NAME;
 import static io.microsphere.i18n.spring.constants.I18nConstants.SOURCES_PROPERTY_NAME;
+import static io.microsphere.i18n.spring.validation.beanvalidation.I18nLocalValidatorFactoryBeanPostProcessor.isValidatorFactoryPresent;
+import static io.microsphere.i18n.spring.web.servlet.AcceptHeaderLocaleResolverBeanPostProcessor.isAcceptHeaderLocaleResolverPresent;
 import static io.microsphere.logging.LoggerFactory.getLogger;
 import static io.microsphere.spring.beans.factory.support.BeanRegistrar.registerBeanDefinition;
 import static io.microsphere.util.ArrayUtils.EMPTY_STRING_ARRAY;
+import static io.microsphere.util.ClassLoaderUtils.getClassLoader;
 import static org.springframework.beans.factory.support.BeanDefinitionBuilder.rootBeanDefinition;
 import static org.springframework.context.support.AbstractApplicationContext.MESSAGE_SOURCE_BEAN_NAME;
 import static org.springframework.core.annotation.AnnotationAttributes.fromMap;
@@ -122,8 +125,13 @@ class I18nImportBeanDefinitionRegistrar implements ImportBeanDefinitionRegistrar
 
     private void registerBeanPostProcessorBeanDefinitions(BeanDefinitionRegistry registry) {
         registerBeanDefinition(registry, ServiceMessageSourceBeanLifecyclePostProcessor.class);
-        registerBeanDefinition(registry, I18nLocalValidatorFactoryBeanPostProcessor.class);
-        registerBeanDefinition(registry, AcceptHeaderLocaleResolverBeanPostProcessor.class);
+        ClassLoader classLoader = getClassLoader(this.getClass());
+        if (isValidatorFactoryPresent(classLoader)) {
+            registerBeanDefinition(registry, I18nLocalValidatorFactoryBeanPostProcessor.class);
+        }
+        if (isAcceptHeaderLocaleResolverPresent(classLoader)) {
+            registerBeanDefinition(registry, AcceptHeaderLocaleResolverBeanPostProcessor.class);
+        }
     }
 
     @Override
