+++ microsphere-i18n-spring/src/test/java/io/microsphere/i18n/spring/beans/factory/config/AcceptHeaderLocaleResolverBeanPostProcessorTest.java
@@ -20,6 +20,7 @@ package io.microsphere.i18n.spring.beans.factory.config;
 
 import io.microsphere.i18n.ServiceMessageSource;
 import io.microsphere.i18n.spring.annotation.EnableI18n;
+import io.microsphere.i18n.spring.web.servlet.AcceptHeaderLocaleResolverBeanPostProcessor;
 import org.junit.jupiter.api.Test;
 import org.springframework.context.annotation.Bean;
 import org.springframework.web.servlet.i18n.AcceptHeaderLocaleResolver;
